requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
selenium==4.15.2
webdriver-manager==4.0.1
pydantic==2.7.0
python-dotenv==1.0.0
apscheduler==3.10.4
beautifulsoup4==4.12.2
requests==2.31.0
pydantic-settings==2.12.0

# Rate limiting
slowapi==0.1.8

# Optional but recommended for production
redis==5.0.1  # For Redis-based rate limiting
lxml==4.9.3  # Faster HTML parsing alternative
httpx==0.26.0  # For async HTTP requests
alembic==1.13.1  # Database migrations
python-multipart==0.0.6  # For file uploads if needed